Moved functions.
Moved non-testing functions from tests.py to utils.py
